pistachio: DDR row/bank/column mapping
The DRAM configuration register, apart from holding the
device density and width also has a rudimentary address
mapping scheme. Currently this is set to the default
Bank/Row/Column. This means that the memory is segmented
into 8 chunks, each with a page detector. If all the
activity is in one section of memory then the other 7
page detectors could be idle.
Changing this to Row/Bank/Column would concatenate the
page detectors meaning that all 8 could be used by a
single initiator. This may not gain anything in a
synthetic bandwidth test but could yield extra performance
in a real world application or benchmark.

BRANCH=none
BUG=chrome-os-partner:37087
TEST=tested on Pistachio bring up board -> DDR initialized
     properly; all access to DDR works properly in
     Coreboot ramstage, Depthcharge and Linux;
     no performance tests were ran so far.

Change-Id: If90b0cf5ce86db5e3d6d362873d22d4269e3a49f
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/264340
Reviewed-by: James Hartley <james.hartley@imgtec.com>
Reviewed-by: David Hendricks <dhendrix@chromium.org>