pistachio: DDR reads return to controller with no bubbles
When the PHY is compiled to run in HDR(half data rate),
then either NOBUB or FXDAT must be set to 1 in the DDR
system general configuration register. NOBUB specifies
that reads should be returned to the controller with
no bubbles and this is felt preferable to the fixed
latency option (FXDAT). Both of them inrease read
latency.

BRANCH=none
BUG=chrome-os-partner:37087
TEST=tested on Pistachio bring up board -> DDR initialized
     properly and ramstage executed correctly

Change-Id: I9ced76bd670fc4efa7441d57e15f97871b046ae9
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/264341
Reviewed-by: James Hartley <james.hartley@imgtec.com>
Reviewed-by: David Hendricks <dhendrix@chromium.org>