auron: use generic broadwell platform.asl
The auron-specific platform.asl is exactly the same except it leaves
out the wake source ACPI methods. This is pretty useless, so we should
just use the broadwell platform.asl instead. This also gets us the
benefit of being able to determine wake sources in the kernel with
no other modifications.

BRANCH=ToT
BUG=chromium:476177
TEST=build, execute on yuna:
  $ echo +10 > /sys/class/rtc/rtc0/wakealarm && powerd_dbus_suspend
  wait for wakeup, check /var/log/messages for wakeup source 00:06

Signed-off-by: Eric Caruso <ejcaruso@chromium.org>
Change-Id: I33e3d49166ccef9d67e7c6a4b00c8c9c6d0f2eeb
Reviewed-on: https://chromium-review.googlesource.com/265335
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>