cbmem: add and use a function to dump console buffer
The new function can be compiled in only when serial console is
disabled.

When invoked, this function initializes the serial interface and dumps
the contents of the CBMEM console buffer to serial output.

Some .h files' tweaking was required to ensure proper exposure of
external declarations of uart_init().

BRANCH=none
BUG=chromium:475347
TEST=compiled for different platforms with and without serial console
     enabled. No actual test of this function yet.

Change-Id: Ib85759a2727e31ba1ca21da7e6c346e434f83b52
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/265293
Reviewed-by: Aaron Durbin <adurbin@chromium.org>