drivers/intel/fsp: pass FSP_INFO_HEADER from original finder
The FSP_INFO_HEADER is located by the assembly code early
within romstage. Use this location as the one source of the
FSP_INFO_HEADER address. To do that a FSP_INFO_HEADER pointer
was added to rosmtage params and those values are passed
through to the users. That way, a find does not need to be
done again.

BUG=chrome-os-partner:38948
BRANCH=None
TEST=None

Change-Id: I95b9b92e61f186967c0e0c2d77a99733e8a967a3
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/265565
Tested-by: Leroy P Leahy <leroy.p.leahy@intel.com>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>