intel/fsp: add FSP binary relocation
When wanting to deploy a RO and updateable RW firmware
scheme when using FSP it's helpful to relocate the FSP binary
into RAM because 1. it's typically linked at a fixed address that
may not match its location in the flash. 2. there's no way to protect
the FSP binary from corruption without saving in a protected memory
region.

This patch does both of those things with help of the SoC support
code. Currently the SoC specific bits for saving and restoring
the support code is not filled out.

The FSP needs to be relocated as a whole Firmware Volume because the
FSP interface does not indicate which programs are utilized from
the FSP_INFO_HEADER. The FSP_INFO_HEADER only contains the the offsets
into the FSP API's entry points. Therefore, every TE image within the
Firmware Volume needs to be relocated on the off chance that the program
may be used by the FSP API calls.

The boot flow changes are that FSP is always relocated and ran from
memory for the silicon init callback as well as the notify steps. In
addition, the silicon init callback is moved from romstage to ramstage.

BUG=chrome-os-partner:38948
BRANCH=None
TEST=Tested logic in standalone program. Needs integration testing.

Change-Id: I668545239cb2ef2d89a29b5b2197f4bde4caba8d
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/265568
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>