Braswell: remove redundant scc gnvs update code from southcluster.c
This is to address issue 38741. Also remove the unneeded register
settings in scc.c due to FSP's done it.
Note that with this change, emmc and sd card are not booting due
to emmc and sd acpi mode are not enabled in devicetree for boards.
It worked earlier because the code in southcluster.c, which is
removed in this patch, regardlessly enables it in gnvs. Enabling
emmc/sd acpi mode in devicetree for cyan and strago are
in separate patches.

BRANCH=none
BUG=chrome-os-partner:38741
TEST=Build and boot to kernel on USB on Cyan
Signed-off-by: robbie zhang <robbie.zhang@intel.com>

Change-Id: I17c0bd27a15da51949eaaaff7492f07236320d93
Reviewed-on: https://chromium-review.googlesource.com/266257
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Robbie Zhang <robbie.zhang@intel.com>
Tested-by: Robbie Zhang <robbie.zhang@intel.com>