Smaug: audio: Set CLK1/MCLK to 12MHz
EXTPERIPH1 aka CLK1 is used to drive AUD_MCLK to
the RT5677 codec @ 12MHz.  Also set DPAUX_SPARE
to 0 to power-up the I2C6 pads.  With this change,
audio beep is heard w/Ctrl+U.

BUG=chrome-os-partner:38343
BRANCH=none
TEST=Builds OK for Smaug and Foster. Ctrl+U at
serial console generates a beep on speakers and
headphones with the serial driver CL in depthcharge.

Change-Id: I28e21b347de84b3f8bb8283605740cad813cd1cb
Signed-off-by: Tom Warren <twarren@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/267133
Reviewed-by: Aaron Durbin <adurbin@chromium.org>