t210: set up appropriate CPU status before jump to ATF
We only bring up CPU0 for ATF and leave the CPU power control with PSCI
method. So we don't need to power on the secondary CPUs. Besides that,
the PMC and flow controller will take over the power sequence control.
Manually control the CPU reset signal is unnecessary. Clear them all in
this patch too.

BRANCH=none
BUG=chrome-os-partner:40414
TEST=CPU can be powered up successful with PSCI method

Change-Id: I6df83d1b5568801040d8ab0ad88a805dc2220aec
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/272415
Reviewed-by: Varun Wadekar <vwadekar@nvidia.com>
Reviewed-by: Tom Warren <twarren@nvidia.com>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>