skylake: Add PCH UART driver for romstage and ramstage
In order to use the PCH UART as a standard serial port it needs
to be initialized properly and put into 16550 (byte address) mode.

The ramstage driver overrides the resource allocator to set the
configured base address so it remains constant.  This could be
changed to update the serial driver with the new base address.

BUG=chrome-os-partner:40857
BRANCH=none
TEST=boot on glados with serial output

Change-Id: I517bb4972aa7cfe0b373c7268ce7cd8571291fd2
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/275159
Reviewed-by: Aaron Durbin <adurbin@chromium.org>