skylake: Clean up and fix device ID handling
The platform information is not being reported properly because
the tables are not filled out properly.  Clean this up and ensure
that the current glados hardware is at least properly identified.

Also add the full list of LPC device IDs as it was only currently
supporting the engineering sample.

BUG=chrome-os-partner:40635
BRANCH=none
TEST=build and boot on glados, see proper platform report early
in the boot log and a lot less scary warnings in the kernel log.

Change-Id: I0e12eca0ea244ae34200851969290829180a8471
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/276343
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Aaron Durbin <adurbin@chromium.org>